requirements.txt
flask==2.3.2
easyocr==1.7.1
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.8.0.76
pillow==10.0.0
numpy==1.24.3
textblob==0.17.1
